requirements.txt
pandas==1.5.3
numpy==1.24.2
matplotlib==3.6.2
pyarrow==10.0.1
Pillow==9.4.0
torch==2.3.1
scikit-learn==1.2.0
seaborn==0.12.1
torchvision==0.18.1
